+++ test/kml.js
@@ -1,7 +1,6 @@
 const test = require('tape')
 const Ogr = require('../src/lib/ogr.js')
 const Helper = require('./helper')
-const _ = require('highland')
 
 test('Set up', t => {
   Helper.before()
@@ -14,28 +13,24 @@ test('Create a kml readstream', t => {
   const rows = []
   Helper.vrt()
   .pipe(Ogr.createStream('kml', options))
-  .on('error', e => {console.log(e); t.end(e)})
+  .on('error', e => { console.log(e); t.end(e) })
   .splitBy('<Placemark>')
   .compact()
-  .each(row => {
-    rows.push(row)
-    })
-  .done(() => {
-    t.equal(rows.length, 101, 'All rows written to the stream')
-   })
+  .each(row => rows.push(row))
+  .done(() => t.equal(rows.length, 101, 'All rows written to the stream'))
 })
 
 test('Gracefully handle a malformed VRT', t => {
   t.plan(1)
   const options = defaultOptions()
   try {
-    _(['Foobar!'])
+    Helper.malformedVrt()
     .pipe(Ogr.createStream('kml', options))
     .on('error', err => {
       t.ok(err, 'Error was caught in the correct place')
       t.end()
     })
-    .done(() => t.end())
+    .on('finish', () => t.end())
   } catch (e) {
     t.fail('Error was uncaught')
   }