+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+* `GeoXForm.createStream` will emit an error when OGR does
+
 ## [1.0.6] - 2016-01-10
 ### Fixed
 * Add missing package